Crate flowrex

flowrex is the minimal executor of flow jobs. It loads a native version of ‘flowstdlib’ flow library to allow execution of jobs using functions provided by ‘flowstdlib’, but it does not load ‘context’ and hence will not execute any jobs interacting with the context.

Modules

errors

We’ll put our errors in an errors module, and other modules in this crate will use crate::errors::*; to get access to everything error_chain creates.

Functions

discover_service 🔒

Try to discover a server offering a particular service by name

get_matches 🔒

main 🔒

Main for flowrex binary - call run() and print any error that results or exit silently if OK

num_threads 🔒

run 🔒

start_executors 🔒